+++ src/parser/inline/extension/indexer.ts
@@ -3,7 +3,6 @@ import { union, creator, surround } from '../../../combinator';
 import { index } from './index';
 import { str } from '../../source';
 import { html } from 'typed-dom';
-import { unshift } from 'spica/array';
 
 export const indexer: ExtensionParser.IndexerParser = creator(surround(
   str(/^\s+(?=\[#[^\s\]])/),
@@ -12,8 +11,4 @@ export const indexer: ExtensionParser.IndexerParser = creator(surround(
   ([, [el]], rest) => [
     [html('span', { class: 'indexer', 'data-index': el.getAttribute('href')!.slice(el.hash.indexOf(':') + 1) })],
     rest
-  ],
-  ([as, bs], rest, context) =>
-    context.syntax?.inline?.extension ?? true
-      ? [unshift<HTMLElement | string>(as, bs), rest]
-      : void 0));
+  ]));